// src/components/models/products.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 崩坏晓店
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Products
 */
export interface Products {
    /**
     * 
     * @type {number}
     * @memberof Products
     */
    'id'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'productName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'description'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Products
     */
    'price'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Products
     */
    'stockQuantity'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'imageurl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'status'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Products
     */
    'categoryId'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Products
     */
    'usersId'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof Products
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'createTime'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Products
     */
    'updateTime'?: string | null;
    'imageurlcopy'?: string | null;
}